Engine: fix bug when caching from process with nested outputs (#5538)
When the execution of a process is being skipped because it has a valid
cache source and caching is enabled, the process would except if the
process has nested output namespaces. Since the nesting cannot be
literally preserved in the database, the nesting is representing by
using double underscores between each level of nesting in the link
label.

The problem occurs when `Process._create_and_setup_db_record` would be
called which would create the outputs of the cloned node, however, it
would use these "collapsed" link labels containing the double
underscores. Passing these to `self.out` would cause the validation to
fail because the process port validation expects the actual nesting.

The solution is to restore the database nesting representation to the
one expected by `Process.out`, which amounts to replacing the namespace
separators.

Cherry-pick: b936da373499c63f4af4bcb9024500ffc9bf7f9e