`DynamicEntryPointCommandGroup`: make actual command dynamic
The `DynamicEntryPointCommandGroup` was recently added to automatically
generate CLI commands for various entry groups in an entry point group.
The options of the generated commands would be dynamically created based
on the class registered with the entry points, however, the actual
command that would be executed was statically defined.

The command simply would create an instance of the class registered at
the entry point, but this might not always be the necessary action. To
make this configurable, the actual command can now be specified when
creating the command group. The only restriction is that it should
accept the loaded entry point as the first argument and kwargs in order
to accept the parsed command line parameters.