`Process`: refactor out input handling specific to `CalcJob` (#5539)
The `CalcJob` subclass of `Process` adds inputs that are specific to its
implementation, namely `code`, `metadata.computer` and `metadata.options`.
Nevertheless, these were being handled in the `Process` class making
assumptions about the exact type of the `code` input.

The code is refactored to move the `CalcJob` specific logic to its
implementation. The signature of `_setup_metadata` had to be changed to
pass in a copy of the metadata inputs as the subclass needs to pop its
custom keys such that `Process` won't raise on keys it doesn't recognize.

The `_setup_inputs` implementation on `CalcJob` is simplified as it now
just needs to set the `Computer` of the `Code` if one hasn't already
been set in the `_setup_metadata` which is called before it.