CLI: normalize use of colors (#5547)
Certain CLI commands used coloring in the output that was inconsistent
with what users might expect based on other parts of the CLI and basic
intuition. For example, red was sometimes used to highlight a purely
informative piece of data or a header, but red is typically associated
with errors.

To normalize the use of colors, all uses of it use the `COLORS` mapping
from the `aiida.cmdline.utils.echo` module and instead of naming the
literal color, it designates the intended meaning.