ORM: raise when trying to pickle instance of `Entity`
The ORM entities should not be pickled since it is not clear whether the
connection to the storage backend should be included as well and if it
were, implementing that in a consistent way across implementations is
not feasible.

Therefore, we now raise an explicit an `InvalidOperation` when an
instance of `Entity` is being pickled. This is achieved by overriding
the `__getstate__` dunder method which will be called by the `pickle`
module when dumping a class instance.