Refactor: remove unnecessary use of `tempfile.mkdtemp`
When it is used for unit tests, it should be replaced by the `tmp_path`
fixture provided by `pytest` since it will take care of cleanup
automatically.

Then there are a few cases in the code where it was used where it was
replaced with `tempfile.TemporaryDirectory`. The latter has the advantage
that it will be automatically cleanup when the object goes out of scope.

A few places in the code still use `mkdtemp` since it needs to control
the cleanup manually.

The `temp_dir` fixture is also deprecated. It works exactly as the
`tmp_path` fixture which ships with `pytest` and should be used instead.