Ensure use of escaping character '\'
The '\' is used for escaping special characters such as '*' and '_'.
This is hard-coded in QueryBuilder but not passed on when constructing
SQL clause, hence replying on the database using '\' for escaping
by default.
For PostgreSQL this works, but not for SQLite where the ESCAPE clause
is needed. Now, we enforce the existence of ESCAPE in all cases.