🐛 FIX: SQLite: apply escape `\` in QueryBuilder `like` and `ilike` (#5553)
For `like` and `ilike`,
the special characters `_` and `%` may need to be escaped if they have literal meanings (`LIKE .. ESCAPE` in SQL).
However, the SQLite backends do not implicitly treat `\` as escape sequence (unlike Postgres): https://sqlite.org/lang_expr.html

This commit makes `\` explicitly an escaping character for SQLite backends. Note, the use of `\` for escaping is already hard-coded in aiida-core.