Add `Complex` data type and support for `numbers.Complex` in database
Serialization/Deserialization of complex numbers is added to the the
psql_dos and sqlite_zip backend. These are stored as objects of the form
`{'__complex__': True, 'real': x.real, 'imag': x.imag}` with the
`__complex__` key being used as the identifier for detecting these
objects for deserializations.

This transformation is not done on the level of `clean_value` but by
customizing the JSON encoder/decoder of the backend. Since `clean_value`
can be called multiple times on the same value thre is no clean way
to distinguish complex numbers that were already transformed or custom
dictionaries, which contain the `__complex__` key (For these a
ValidationError is raised).

Added support for complex attributes in QueryBuilder
complex numbers are compared like objects. For comparisons
attribute.complex.real/imag can be used

Open Questions:
  - What are the performance implications of customising the json
    serialization/deserialization
  - Should the Jsonable class handle complex numbers?