🐛 FIX: Archive creation after packing repository (#5570)
`disk_objectstore.PackedObjectReader` does not implement the ability to `SEEK_END`,
required to determine the size of a stream.
The size of the stream is required to determine if the ZIP64 extension should be used,
so in this case, we now defer to `force_zip64=True`.

Cherry-pick: 46d244e32ac5eca2e22a3d088314591ce064be57