RestAPI: CustomJSONEncoder support bytes objects
In Aiida 2.0, REST (where flask 1 was updated to flask 2), bytes objects are not supported anymore.
Because Flask 1.x was based on simplejson, that inherently converts bytes objects into strings. 
In flask 2.0.0, it was redesigned based on the default python json. 
See here: https://flask.palletsprojects.com/en/2.0.x/changes/ and here pallets/flask#3555
The conversion of a bytes array into `str` in the `CustomJSONEncoder` is implemented to workaround. 

Co-authored-by: Jusong Yu <jusong.yeu@gmail.com>