RESTAPI: treat 'false' as False in url parsing (#5573)
fixes #3635

The parser function that we use to parse the query string doesn't parse booleans as expected.
When passing `false` in URL it will pared as `True` which is not correct.
This PR fixes it by comparing the `true` or `false` attribute in url directly with the string `'true'` and return bool type.

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>

Cherry-pick: 7772adcae4113404eef27e1e964d7d55734c924c