`BaseRestartWorkChain`: fix `handler_overrides` ignoring `enabled=False`
In commit `7b8c61d0869751455993891cfe5cf6c637593344`, the input that
allows overriding handlers, `handler_overrides`, was updated to include
the possibility of overriding the priority. The changes broke the
original behavior though where a handler that was disabled by default
could not be enabled by specifying `enabled=False` in the overrides.