Add global config option `rest_api.profile_switching`
This global config option is set to `False` by default. When set to
`True`, the REST API will allow requests to specify the profile and it
will switch the loaded profile when necessary. If a request specifies
the profile query parameter and profile switching is turned off, a 400
Bad Request response is returned.