Add Complex data type and support for numbers.Complex in database
Alternative implementation of #5561

These are stored as objects of the form
`{'__complex__': True, 'real': x.real, 'imag': x.imag}` with the
`__complex__` key being used as the identifier for detecting these
objects for deserializations.

For serialization/deserialization we use clean_value and a new function
deserialize_value. `clean_value` is pulled out of BackendNode and is
now done on the level of NodeAttributes and NodeExtras. The
deserialization is added to the get/get_many and all routines of these
classe

Added support for complex attributes in QueryBuilder
complex numbers are compared like objects. For comparisons
attribute.complex.real/imag can be used

Open Questions:
  - What are the performance implications of the added deserialization?
    (Since clean_value is already called I think there would be no impact
     from that side)
  - Jsonable class should handle complex numbers. Setting attributes
    work, but the initial check json.loads(json.dumps(...)) fails