DevOps: Add a notification for nightly workflow on fail (#5605)
When the `nightly.yml` workflow fails it will now send a notification to
the `dev-aiida-core` channel on the AiiDA Slack.

The specific logic in the conditional is on purpose:

    always() && (steps.install.outcome == 'Failure' || steps.tests.outcome == 'Failure')

The reason is that without the `always()`, the step will never run if
any of the previous steps fail, defeating the purpose. However, this
is not enough, because with just `always()` the final status of the
workflow will be that of the notification step and so will always be
successful. This is confusing since a failed build will be marked as
successful.

By adding the clause after `&&` the build will still be marked as failed
even if the final notification step runs fine.