Dependencies: set upper limit `werkzeug<2.2` (#5606)
The `werkzeug==2.2` release is breaking the REST API causing most of the
routes to return a 404 error. The problem is that routes with a trailing
slash are now treated as branch leaves, whereas before they would be
treated as leaf routes if that could be uniquely determined. Now it will always
treat it literally and return a 404 if not explicitly defined.