`CalcJob`: Add functionality that allows live monitoring
Instead of killing the process, it is preferable to have the engine
retrieve the files from the job and call the parser, if one was defined.
However, this would result in the process to be marked as terminated
nominally with the exit status returned by the parser. It would be
impossible to see that the job was stopped by a monitor other than from
the process report. It is important that one can query for processes
that were stopped by a monitor. To do so, a dedicated exit code is
defined on the `CalcJob` called `STOPPED_BY_MONITOR` which is set on the
node, overriding any exit status returned by the parser. There is no way
around this since only one exit status can be set on a node.
Nevertheless the result from the parser is logged and so is visible in
the process report.

If monitors are defined for the `CalcJob` the corresponding node will
contains information of the package versions from which the monitors
come. The information is added to the `version` attribute which already
contains version information of `aiida-core` and the `CalcJob` plugin
itself.