Archive: add missing migration of transport entry points (#5604)
In the migration from version `0.11` to `0.12` the existing entry points
from `aiida-core` were updated to include the `core.` prefix, however,
the entry points of transports were forgotten. Since the incorrect
migration was already released with `v2.0` of `aiida-core`, a separate
migration has to be added to perform the transport entry point migrations.