Define `shell_type` in `Transport` and `JobTemplate`
The `_shell_type` attribute of a `Transport` class represents the type
of shell used by the remote computer, e.g. `bash` (default) or
`powershell`, and should be defined within the subclasses of
`Transport`. This will be the command used to execute commands through
transport.
A `shell_type` field has also been added to `JobTemplate`. This is set
by `CalcJob.presubmit` to match the `shell_type` property of the
computer's transport. The `shell_type` property can be also used by a
direct scheduler to directly execute commands, using the computer's
specific shell.