`BaseRestartWorkChain`: add the `get_outputs` hook
The `BaseRestartWorkChain`, in the `results` outline step, will take the
outputs of the last completed `CalcJob` and attach them to the work
chain itself. There are use cases where the implementation may want to
slightly adapt the outputs before attaching them. By abstracting the
code that fetches the outputs to a separate method `get_outputs`, it
provides a hook for implementations to customize this behavior.