`ArrayData`: replace `nan` and `inf` with `None` when dumping to JSON (#5613)
The `ArrayData` data type stores data as numpy arrays, which supports
the `nan` and `inf` special float values in Python. However, these are
not supported by the JSON standard. This poses a problem when the array
node needs to be serialized to JSON, for example when it is served over
the REST API.

The solution is to convert all `nan` and `inf` values to `None`, which
will be converted to `null` when serialized to JSON. This now makes it
possible to serve `ArrayData` nodes over the REST API.