API: Add the `aiida.engine.processes.control` module
This module abstracts functionality that was implemented directly in the
`aiida.cmdline.commands.cmd_process` module that allows to interact with
active processes. By extracting it to a separate module, this
functionality can now also be easily used by users from interfaces other
than the CLI, such as the shell, or a notebook.