Fixtures: Fix the `with_daemon` test fixture
The `with_daemon` fixture would not properly shut the daemon process
down. This is addressed by using the new API on the `DaemonClient` to
start and stop the daemon. This simplifies the code significantly.

The `with_daemon` fixture will ensure that the daemon is running by
getting a daemon client configured for the profile of the current test
session and starting it if it isn't already. It then yields the client
to the test, so that the test could even stop the daemon if needed for
the test. The `daemon_client` fixture that provides the client, is
session scoped and will make sure that the daemon is stopped at the end
of the test session.

This new implementation of starting the daemon through `DaemonClient`
exposed a bug in the `aiida.manage.configuration.profile` module. It
surfaces when running the test with a temporary profile that is
generated on the fly by the `aiida.manage.tests.main.test_manager`
context manager. In this case, a completely isolated AiiDA configuration
is created in a temporary directory. However, these changes are not
reflected in the various constant variables that are declared at the top
of the `profile` module. This includes the path to the daemon PID for
example, and this will cause the stopping of the daemon to fail since it
is looking for the PID file in the wrong location and so cannot find it.

The solution is to make the definition of the daemon filepaths dynamic
by having them generated within the `Profile.filepaths` property
including the import of the `aiida.manage.configuration.settings`
directory. This ensures that the test manager infrastructure can
temporarily set the `AIIDA_PATH` environment variable to the path of the
temporary directory for the test session and call the function
`set_configuration_directory` of the `settings` module ensuring that the
right filepaths are set for the various daemon files.