`Profile`: make definition of daemon filepaths dynamic
The filepaths relevant to the daemon that are provided through the
`Profile.filepaths` property, where defined statically at the module
level. This means that they would be fixed as soon as the `profile`
module would be first imported based on the `DAEMON_DIR` and the
`DAEMON_LOG_DIR` variables of the `settings` module.

This would be problematic if those variables were to be changed, because
they wouldn't be reflected by `Profile.filepaths`. An example of such a
case is the running of the test suite without a pre-configured test
profile. In this case, a temporary test profile is created from scratch
including the AiiDA configuration directory. Since the path of the
configuration directory changes, the filepaths of the daemon also
changed and this would have to be reflected by `Profile.filepaths`.