`DaemonClient`: add the `get_env` method
The `DaemonClient.get_env` method is added which provides a copy of the
environment that is used by the process in which the client is running.
This is used by the `start_daemon` and `_start_daemon` methods to ensure
that the launched subprocesses use the exact same environment as the
process of the client.

The implementation is largely taken from the `get_env_with_venv_bin`
function of the `aiida.cmdline.utils.common` utility. This function is
deprecated since its is purely relevant to the daemon functionality and
so should just be contained in the `DaemonClient`.

There is a single important change and that is that `get_env` adds the
current content of `sys.path` to the `PYTHONPATH` key. This is important
as to guarantee that the daemon process will have the same module paths
as the main process. This ensures that all modules that are importable
by the main process, will also be importable by the daemon process.