Fixtures: Add `started_daemon_client` and `stopped_daemon_client`
The fixtures `started_daemon_client` and `stopped_daemon_client` are
added which are function scoped fixtures that ensure that the daemon is
running or stopped, respectively. The fixtures return an instance of the
`DaemonClient` which can be used to interact further with the daemon
where necessary.

The fixtures both use the `daemon_client` fixture which is session
scoped and will ensure that the daemon is stopped at the end of the test
session. The `started_daemon_client` replaces the `with_daemon` fixture
which was not going through the `DaemonClient` designed for this
purpose.