`Repository.copy_tree`: omit subdirectories from `path` when copying
The intention for `copy_tree` was to copy only the contents of the
relative subpath, however, it was including the subdirectories. This was
unintentional and users also probably won't expect this behavior. For
example, when copying the content of the `relative/sub` subdirectory,
one wouldn't expect the content to be copied inside a `relative/sub`
folder as well. This is also not the behavior of `shutil.copytree`.
When executing `shutil.copytree('relative/sub', target`), the folder
`relative/sub` won't be present in the `target` directory.

The solution is to make the target path of the file to be copied
relative to the source path within the repository.

Although this change is technically breaking backwards compatibility, it
is correcting a bug of a feature that was introduced in `v2.0`. It is
very unlikely that any plugins started depending on this behavior so we
still have the possibility to fix it now.

Cherry-pick: 7a39ad21256d3ad8ca156781277b92c182da0801