`CalcJob`: fix bug in `local_copy_list` provenance exclusion
When the `local_copy_list` was provided a directory as source, the
contents would be copied to the sandbox folder in `upload_calculation`.
The point of using the `local_copy_list` is to guarantee that the files
copied this way, are not also copied to the repository of the process
node, since this would duplicate content. This was accomplished by
adding the files copied through this method to `provenance_exclude_list`.

The implementation contained a bug though when the `local_copy_list`
instruction also defined a target. In this case the target was not
included in the paths to be excluded and so were missed when copying the
contents of the sandbox to the repository of the process node. This is
corrected and regression tests are added.

Cherry-pick: 0bfe0008e27076fffba8dbcc4d3883509e1e4401