`JobResource`: add the `accepts_default_memory_per_machine` (#5642)
In ff1318b485a8b803e115b78946cc4593fc661153 the `Computer` class got the
`get_default_memory_per_machine` method which returns the default memory
to request from the scheduler when submitting a calcjob to this machine.

The `verdi computer setup` interface got an additional option to set this
value. However, for certain computer setups, for example those defining
the `DirectScheduler` as the scheduler type, the memory should or cannot
typically be defined. After the CLI started asking for this value by
default, jobs submitted to a computer with the `core.direct` scheduler
started emitting warnings as the default memory would be passed, which
would subsequently be ignored.

The solution is to extend the `JobResource` class with the property
`accepts_default_memory_per_machine`. This returns `True` by default,
keeping the current behavior, but it can now be overridden by the
`DirectScheduler` to return `False`. The CLI is updated to only prompt
for the default minimum memory for a computer if the scheduler type
return `True` for this property.