`CalcJobMonitor`: Add the `priority` attribute
The `priority` attribute takes an integer and is zero by default. It
allows the user to define the order in which monitors need to be called
in case multiple are defined.

The ordering is implemented in the `CalcJobMonitors` utility class. This
is mostly so that it is easy to unit test. It ensures the monitors are
ordered by their priority, going from high to low. In case of identical
priorities, the monitors are sorted alphabetically by the keys in the
imonitors` input namespace.