`CalcJobMonitor`: Add the `minimum_poll_interval` attribute
The `minimum_poll_interval` is an optional positive integer that can be
defined for a monitor. If defined, the engine will ensure that the
interval between two successive calls of the monitor will at least be
this long.

In order to track the last time a monitor was called, the timestamp is
added to the `call_timestamp` attribute when it is called by the
`CalcJobMonitors.process` method.