Add the `CalcJobMonitorResult` dataclass
This dataclass can be returned by a monitor to communicate to the engine
the course of action to take. Returning a str from a monitor remains
supported as it is automatically converted to a `CalcJobMonitorResult`.

The first attribute that is added is `action` which takes an instance of
the `CalcJobMonitorAction` enum. Currently the only value is `KILL`
which is therefore also the default and instructs the engine to kill the
job and stop monitoring.