`ArithmeticAddCalculation`: Add the `metadata.options.sleep` input
When set, this will add a sleep in front of the main command. This is
useful for testing where the calculation needs to run at least for a
certain time in order for another component to get the chance to run. An
example is to test `CalcJob` monitors. Without the sleep, the job would
finish before they get the chance to analyze the output files.