`CalcJobMonitorResult`: Add the `override_exit_code` attribute
By default it is `True`, but if set to `False`, the engine will not
override the exit code returned by the parser with the default exit code
`STOPPED_BY_MONITOR` that is set when a job is stopped through a monitor.
Naturally, this attribute is ignored when the `parse` and or `retrieve`
attribute are set to `False` as in that case the parser is never even
called so there is nothing to not override.