`CalcJobMonitorAction`: Add the `DISABLE_ALL` option
So far the `CalcJobMonitorAction`, the enum that can be returned by a
calcjob monitor, only supported a single option `KILL`. When the engine
receives this instruction, the job will be killed immediately.

An alternative use case is where a monitor has performed a check and an
optional action and now simply wants to let the job run its course.
Often in this case it is important that the monitor itself, and any
others that may have been registered, are no longer run for the lifetime
of the job.

The `CalcJobMonitorAction` now adds the `DISABLE_ALL` option, which when
set on the `action` attribute of a `CalcJobMonitorResult`, the engine
will disable all monitors for the remainder of the duration of the job.