`CalcJobMonitorAction`: Add the `DISABLE_SELF` option
The default action for a `CalcJobMonitorResult` is the option
`CalcJobMonitorAction.KILL` which immediately kills the job. However,
sometimes, one wants to simply disable the monitor and continue running
the job nominally. The `CalcJobMonitorAction.DISABLE_SELF` option
instructs the engine to not call the monitor that returned it again in
future monitor evaluations.