`AbstractCode`: add abstraction methods for command line parameters (#5664)
This should be the final step in the refactoring of `CalcJob.presubmit`
that will make it fully independent of the actual subclass that is used
for the code input to the calculation job.

Methods `get_executable_cmdline_params` and `get_prepend_cmdline_params`
are added to the `AbstractCode` class. The former receives a list of
command line argument that the `CalcJob` plugin has provided. The code's
job is to format them in any particular way if necessary and prepend the
the code's executable.

The latter should format the command line parameters that will be pre-
pended to the executable. This is typically used for MPI parameters.
These are communicated through the `mpi_args` and `extra_mpirun_params`
arguments which represent the arguments provided by the
`Computer.get_mpirun_command` and the `mpirun_extra_params` metadata
option input.

Co-authored-by: Sebastiaan Huber <mail@sphuber.net>