`TemplatereplacerDoublerParser`: rename and generalize implementation (#5669)
The `TemplatereplacerDoublerParser` is a parser dedicated to be used in
combination with the `TemplatereplacerCalculation`, but for a very
specific code that simply doubles a number using bash. The parser is
most likely only used by `aiida-core` for testing purposes, but it can
be made more useful by being slightly generalized.

It is renamed to `TemplatereplacerParser` and instead of always casting
the output written to stdout to an integer, the output is simply kept as
is. This makes the parser usable for a larger number of use-cases.