Docs: Remove explicit release tag in Docker image name (#5671)
The documentation on running AiiDA in a Docker container was suggesting:

    docker pull aiidateam/aiida-core:|release|

Where `|release|` would be substituted with the current version of the
package that was being built. However, since the v2.0 release, we started
adding the `.post0` suffix to the version number in the `main` branch.
This means that the `latest` documentation branch, which builds directly
of the `main` branch would generate, for example, the instruction:

    docker pull aiidateam/aiida-core:2.0.4.post0

But this would of course fail as that version does not exist on Docker
hub. Instead, we can simply use the `latest` identifier:

    docker pull aiidateam/aiida-core:latest

Since the `latest` build of the documentation should reference the latest
code release, this should be normally be correct.