DevOps: Fix the nightly slack notification on PRs from forks
A limitation of secrets in Github Actions is that they are not available
in workflow runs that are triggered by pull requests that come from
forks. This would cause problems for the nightly workflow, which when
failed tried to send a notification to the Slack channel. However, this
requires the `SLACK_WEBHOOK` which is not available and so the step
would error out.

The solution is to only execute the step if the secret is available
which is now explicitly checked for in the conditional.