Tests: Update entry point string `TemplatereplacerParser` in nightly (#5674)
In commit 3416ba6afb52ed72eec7883e41d1c56ea5a6670d, the entry point name
of the `TemplatereplacerDoublerParser` was renamed to
`core.templatereplacer` but the daemon tests were not updated, causing
the nightly build to fail.

Since the `TemplatereplacerDoublerParser` was generalized and renamed to
`TemplatereplacerParser`, the results parsed from the doubling code are
no longer converted to an integer by the parser, but insted the result
will simply be the exact content parsed from the file, e.g., `10\n`. The
`validate_calculations` method is updated to strip the newline and
cast the value to an integer to make the comparison pass.