Tests: Add the `EntryPointManager` exposed as `entry_points` fixture (#5656)
This fixture allows to easily add or remove entry points for the scope
of the function. It temporarily manipulates the content of the entry
point cache as provided by `importlib.metadata.entry_points`. After the
test function ends, the additions and removals are undone.

The `SelectableEntryPoints` class fully implements the dictionary
interface which we use to update its state, however, this interface is
deprecated. The methods that replace it though, the `select` method,
are read-only and do not allow to modify the state, so we cannot use
them. To not flood the logs with deprecation warnings, they are
intentionally suppressed.