CLI: Add the `verdi devel revive` command
This method exposed the newly added `revive_processes` method from the
process control module. It is put under `verdi devel` since this command
should be used prudently and as a last resort, when the cause of the
problem is known with high certainty, because incorrect use of the
command can create bigger problems.