`verdi setup`: Add the `--profile-uuid` option (#5673)
When setting up a profile in multiple AiiDA instances that should all
use the same services (storage backend, message broker), for this to
work, the profile UUID should be identical. Currently, there is no way
to set the UUID of a profile through the CLI. The `verdi setup` and
`verdi quicksetup` commands will always generate a new random UUID.

Here we add the `--profile-uuid` option to `verdi setup` which allows to
specify the desired UUID instead of having one generated. It is not
added to `verdi quicksetup` because in this use case of sharing
resources between profiles, the resource is normally created before
hand. Likewise, this use case is often deployed automatically which uses
`verdi setup` in a non-interactive way. This is why the option will not
be prompted for in interactive mode. That would only confuse normal
users that are setting up a normal profile where the UUID should simply
be generated automatically.