Process control: Add the `revive_processes` method
There is a known bug where the process task to continue the execution of
a `Process` that is sent to RabbitMQ gets lost before the process is
completed. This will result in the process never getting picked up
again by the daemon and it will seem "stuck".

If this is the case, there is a simple remedy and that is to simply
recreate the process task and send it to RabbitMQ. The `revive_processes`
method is added to provide a simple interface for this action.

Note that this action should be used sparingly and as a last resort when
all measures have been taken to be as sure as possible that the reason
for the process being stuck is actually because the task got lost.
Unfortunately, due to the model of RabbitMQ, the contents of the process
queue cannot be inspected as consuming the messages is destructive and
they are lost upon reading them.