`ProcessFunction`: Automatically serialize Python base type inputs
The `to_aiida_type` dispatcher is set as the `serializer` of the input
ports of the `ProcessFunction` process class. This process class is
generated on the fly for a function decorated by the `calcfunction` and
`workfunction` decorators.

Through change, process functions can now be called with Python base
types for their arguments, which will be automatically wrapped in the
corresponding AiiDA data type.