Tests: Suppress RabbitMQ and developer version warnings
Warnings are raised when a profile is loaded that configures a RabbitMQ
server with an unsupported version or if the installed `aiida-core` code
is not a released version. These warnings are not relevant for testing
and so they are suppressed by setting the relevant config options.

The options are set on the automatically created config in the case of
the temporary test profile, as well as the test profile that is created
manually before for the tests run in the Github Actions workflow.